refactor: Move all activities into ui folder for better project structure
- Created a new folder named 'ui'
- Moved all activities into the 'ui' folder
- Updated package declarations and imports as necessary
