+++ app/src/main/AndroidManifest.xml
@@ -70,7 +70,7 @@
             </intent-filter>
 
     </service>
-        <activity android:name="org.bepass.oblivion.SplashScreenActivity"
+        <activity android:name="org.bepass.oblivion.ui.SplashScreenActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -79,7 +79,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="org.bepass.oblivion.MainActivity"
+            android:name="org.bepass.oblivion.ui.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.OblivionUI">
             <intent-filter>
@@ -89,20 +89,20 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="org.bepass.oblivion.InfoActivity"
+            android:name="org.bepass.oblivion.ui.InfoActivity"
             android:exported="false" />
 
         <activity
-            android:name="org.bepass.oblivion.LogActivity"
+            android:name="org.bepass.oblivion.ui.LogActivity"
             android:exported="false" />
 
         <activity
-            android:name="org.bepass.oblivion.SettingsActivity"
+            android:name="org.bepass.oblivion.ui.SettingsActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustResize" />
 
         <activity
-            android:name=".SplitTunnelActivity"
+            android:name=".ui.SplitTunnelActivity"
             android:exported="false" />
     </application>
