+++ app/src/main/java/org/bepass/oblivion/ui/InfoActivity.java
@@ -1,4 +1,4 @@
-package org.bepass.oblivion;
+package org.bepass.oblivion.ui;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -7,6 +7,8 @@ import android.widget.ImageView;
 import android.widget.RelativeLayout;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import org.bepass.oblivion.R;
 
 public class InfoActivity extends AppCompatActivity {
