+++ app/src/main/java/org/bepass/oblivion/ui/SplitTunnelActivity.java
@@ -1,15 +1,22 @@
-package org.bepass.oblivion;
+package org.bepass.oblivion.ui;
 
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageView;
 
-import androidx.activity.OnBackPressedCallback;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.ConcatAdapter;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.google.android.material.progressindicator.CircularProgressIndicator;
+
+import org.bepass.oblivion.BypassListAppsAdapter;
+import org.bepass.oblivion.ConnectionState;
+import org.bepass.oblivion.FileManager;
+import org.bepass.oblivion.R;
+import org.bepass.oblivion.SplitTunnelMode;
+import org.bepass.oblivion.SplitTunnelOptionsAdapter;
+import org.bepass.oblivion.base.StateAwareBaseActivity;
 
 
 public class SplitTunnelActivity extends StateAwareBaseActivity {
