+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
   android:layout_width="match_parent"
   android:layout_height="match_parent"
   android:background="@drawable/background_gradient"
-  tools:context="org.bepass.oblivion.MainActivity">
+  tools:context="org.bepass.oblivion.ui.MainActivity">
 
   <com.google.android.material.floatingactionbutton.FloatingActionButton
       android:id="@+id/floatingActionButton"
