+++ app/src/main/res/layout/activity_split_tunnel.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/background_gradient"
     android:orientation="vertical"
-    tools:context=".SplitTunnelActivity">
+    tools:context=".ui.SplitTunnelActivity">
 
 
     <RelativeLayout
